+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feat(grafana_cloud_logs): add support for Grafana Cloud Logs ([#92](https://github.com/fastly/terraform-provider-fastly-beta/pull/92))
 - feat(ngwaf_workspace_templated_signal_rules): moved Templated Signal Rules to the dedicated `fastly_ngwaf_workspace_templated_signal_rules` data source ([#90](https://github.com/fastly/terraform-provider-fastly-beta/pull/90))
+- feat(digitalocean_logging): add support for Digital Ocean Logging ([#95](https://github.com/fastly/terraform-provider-fastly-beta/pull/95))
 - feat(user_service_authorization): add `fastly_user_service_authorization` resource for granting user permissions on a service ([#94](https://github.com/fastly/terraform-provider-fastly-beta/pull/94))
 - feat(api_security_discovered_operations): add `fastly_api_security_discovered_operations` data source ([#96](https://github.com/fastly/terraform-provider-fastly-beta/pull/96))